services/forecast-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
redis==5.0.4
apscheduler==3.10.4
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ML
prophet==1.1.5
torch==2.3.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
joblib==1.4.2
xgboost==2.0.3

# Experiment tracking
mlflow==2.13.0

# Testing
ruff==0.4.4
mypy==1.10.0
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
